제약조건 추가 및 변수명 수정

+++ todolist/src/main/java/com/example/demo/controller/TodoListController.java
@@ -12,6 +12,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+
 
 import java.time.LocalDate;
 import java.util.List;
@@ -35,7 +37,7 @@ public class TodoListController {
 
         // 이메일로 사용자 조회
         User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new RuntimeException("해당 사용자를 찾을 수 없습니다."));
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.UNAUTHORIZED, "해당 사용자를 찾을 수 없습니다."));
 
         // 할 일 목록 생성
         ToDoList toDoList = ToDoList.builder()
@@ -58,19 +60,22 @@ public class TodoListController {
 
         // 이메일로 사용자 조회
         User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new RuntimeException("해당 사용자를 찾을 수 없습니다."));
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.UNAUTHORIZED, "해당 사용자를 찾을 수 없습니다."));
 
         // 사용자 ID와 요청 ID로 할 일 항목 조회
-        ToDoList todo = toDoListRepository.findAllByUserIdAndId(user.getId(), request.getId())
-                .orElseThrow(() -> new RuntimeException("해당 할 일을 찾을 수 없습니다."));
 
+        if(toDoListRepository.findAllByUserIdAndId(user.getId(), request.getId()).isPresent()) {
+            ToDoList todo = toDoListRepository.findAllByUserIdAndId(user.getId(), request.getId())
+                    .orElseThrow();
+            todo.setIsCheck(request.getIsCheck());
+            // 저장
+            toDoListRepository.save(todo);
+            return ResponseEntity.ok("체크 작성 성공");
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("글을 찾을 수 없습니다.");
+        }
         // 체크 여부 설정
-        todo.setIsCheck(request.getIsCheck());
 
-        // 저장
-        toDoListRepository.save(todo);
-
-        return ResponseEntity.ok("체크 작성 성공");
     }
 
     //수정
@@ -81,7 +86,7 @@ public class TodoListController {
 
         // 이메일로 사용자 조회
         User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new RuntimeException("해당 사용자를 찾을 수 없습니다."));
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.UNAUTHORIZED, "해당 사용자를 찾을 수 없습니다."));
         try {
             ToDoList todo = toDoListRepository.findAllByUserIdAndId(user.getId(), request.getId())
                     .orElseThrow(() -> new RuntimeException("해당 ID의 할 일이 없습니다."));
@@ -103,7 +108,7 @@ public class TodoListController {
 
         // 이메일로 사용자 조회
         User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new RuntimeException("해당 사용자를 찾을 수 없습니다."));
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.UNAUTHORIZED, "해당 사용자를 찾을 수 없습니다."));
 
         List<ToDoList> todoLists = toDoListRepository.findAllByUserId(user.getId());
 
@@ -124,7 +129,7 @@ public class TodoListController {
 
         // 이메일로 사용자 조회
         User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new RuntimeException("해당 사용자를 찾을 수 없습니다."));
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.UNAUTHORIZED, "해당 사용자를 찾을 수 없습니다."));
 
         LocalDate localDate = LocalDate.parse(date); // YYYY-MM-DD 날짜 별로
         List<ToDoList> todoLists = toDoListRepository.findAllByUserIdAndDoDate(user.getId(),localDate);
@@ -139,15 +144,19 @@ public class TodoListController {
                 .toList();
     }
 
-    //삭제   삭제시 유저 확인 필요
+    //삭제
 
     @DeleteMapping("/{id}")
     public ResponseEntity<String> delete(@RequestHeader("Authorization") String token, @PathVariable Long id) {
         String email = jwtTokenProvider.getUserEmailFromToken(token);
         User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new RuntimeException("해당 사용자를 찾을 수 없습니다."));
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.UNAUTHORIZED, "해당 사용자를 찾을 수 없습니다."));
 
-        todolistService.deleteToDo(user.getId(), id);
-        return ResponseEntity.ok("글 삭제 성공");
+        if(toDoListRepository.findAllByUserIdAndId(user.getId(), id).isPresent()) {
+            todolistService.deleteToDo(user.getId(), id);
+            return ResponseEntity.ok("글 삭제 성공");
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("글을 찾을 수 없습니다.");
+        }
     }
 }
